# src/MooBank.Api/openapi-v1.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "MooBank API",
    "version": "2026.7.0"
  },
  "servers": [
    {
      "url": "/api"
    }
  ],
  "paths": {
...
    "/bills/import": {
      "post": {
        "tags": [
          "Bills"
        ],
        "operationId": "import-bills",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Import"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportResult"
                }
              }
            }
          }
        }
      }
    },
...
    "/instruments/{instrumentId}/accounts/{accountId}/import": {
      "post": {
        "tags": [
          "Import"
        ],
        "operationId": "import",
        "parameters": [
          {
            "name": "instrumentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "accountId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "required": [
                  "file"
                ],
                "type": "object",
                "properties": {
                  "file": {
                    "$ref": "#/components/schemas/IFormFile"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/instruments/{instrumentId}/accounts/{accountId}/import/reprocess": {
      "post": {
        "tags": [
          "Import"
        ],
        "operationId": "reprocess",
        "parameters": [
          {
            "name": "instrumentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "accountId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Discount": {
        "type": "object",
        "properties": {
          "discountPercent": {
            "type": [
              "null",
              "integer"
            ],
            "format": "uint8"
          },
          "discountAmount": {
            "type": [
              "null",
              "number"
            ],
            "format": "double"
          },
          "reason": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "IFormFile": {
        "required": [
          "contentType",
          "contentDisposition",
          "headers",
          "length",
          "name",
          "fileName"
        ],
        "type": "string",
        "format": "binary"
      },
      "Import": {
        "required": [
          "bills"
        ],
        "type": "object",
        "properties": {
          "bills": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ImportBill"
            }
          }
        }
      },
      "ImportBill": {
        "required": [
          "accountName",
          "issueDate",
          "periods",
          "discounts"
        ],
        "type": "object",
        "properties": {
          "accountName": {
            "type": "string"
          },
          "invoiceNumber": {
            "type": [
              "null",
              "string"
            ]
          },
          "issueDate": {
            "type": "string",
            "format": "date"
          },
          "currentReading": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          },
          "previousReading": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          },
          "total": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          },
          "costsIncludeGST": {
            "type": [
              "null",
              "boolean"
            ]
          },
          "cost": {
            "type": [
              "null",
              "number"
            ],
            "format": "double"
          },
          "periods": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Period"
            }
          },
          "discounts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Discount"
            }
          }
        }
      },
...
      "ImportResult": {
        "required": [
          "imported",
          "failed",
          "errors"
        ],
        "type": "object",
        "properties": {
          "imported": {
            "type": "integer",
            "format": "int32"
          },
          "failed": {
            "type": "integer",
            "format": "int32"
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "Period": {
        "required": [
          "periodStart",
          "periodEnd",
          "pricePerUnit",
          "totalUsage",
          "chargePerDay"
        ],
        "type": "object",
        "properties": {
          "periodStart": {
            "type": "string",
            "format": "date-time"
          },
          "periodEnd": {
            "type": "string",
            "format": "date-time"
          },
          "daysInclusive": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          },
          "days": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          },
          "pricePerUnit": {
            "type": "number",
            "format": "double"
          },
          "totalUsage": {
            "type": "number",
            "format": "double"
          },
          "cost": {
            "type": [
              "null",
              "number"
            ],
            "format": "double"
          },
          "chargePerDay": {
            "type": "number",
            "format": "double"
          }
        }
      },
...
    },
    "securitySchemes": {
      "oidc": {
        "type": "openIdConnect",
        "description": "OIDC",
        "openIdConnectUrl": "https://login.microsoftonline.com/30efefb9-9034-4e0c-8c69-17f4578f5924/v2.0/.well-known/openid-configuration"
      }
    }
  },
  "security": [
    {
      "oidc": [ ]
    }
  ],
...
}